move `sized_button` into `ui/widgets` as `styled_button`
